requirements.txt
fastapi
uvicorn
transformers
pydantic
python-dotenv
torch
